src/gemini-file-search/requirements.txt
google-genai>=1.30.0